prevent container membership changes while an array is stopping
This commit fixes an issue where, on a system with slow array updates, the
manager thread could wake up and catch a container array that's half removed,
and start changing the container membership in error.

In this condition, the monitor thread notices that the sub-array has been
stopped, but is bound up synchronizing metadata before it lets the manager
thread know about the change.  In cases where the monitor thread is backed up
on metadata updates, say, due to sync finishing, and a pending write, this can
take some time to work itself out.  This is compounded by the protocol that
mdadm has, where it merely needs the monitor thread to say "yeah, I heard you".
In our system, then heal likely comes along and stops the container too shortly
after.  Sometimes this catches the manager thread just at the right instant to
cause some chaos.

This change allows the monitor thread to raise an "array_stopping" flag as soon
as it sees the array stopping.  The monitor thread avoids making changes to the
container (in manage_container) if this flag has been raised.

As a 2nd-level protection, the manager now checks that the /proc/mdstat output
includes a member sub-array for the container it's managing.  If there's no
active sub-array in there, then the manager won't try to alter the container
membership.

additionally:

- monitor thread logs "ARRAY STOPPED" when it notices that the array has
  stopped
- the monitor thread's "wake" message now uses a single fprintf() call, so it
  can't end up split across output lines