advance the sequence number significantly on assembly + pre-flight
This fix addresses a bug observed at Netsurf, where the admin did a soft
"reboot" to test failover of the active member.  Mdvote was stopped first, and
the system array ended up failing a drive out on the way down.  The essentially
partitioned cm1 made slight forward progress on its local drive, incrementing
its sequence number by one or two.

The cluster failed over to cm2, where it assembled the system array with the
local drive there and properly flushed the minimum assembly sequence number to
mdvote_agent.  It ran like that for a minute before it was rebooted, too.  Both
nodes are down now.

Cm1 came back online and found that its local drive met the minimum assembly
sequence number, due to the progress it made while partitioned.  It assembled
and ran the array, without any of the data written on cm2.

This fix has two parts.

First, it has a pre-flight check.  Before attempting to write the superblocks
of the remaining drives, mdmon now does an HTTP GET to mdvote_agent to check
that it's up.  Any failure here is handled as a retry.  This part of the fix
narrows the window of the observed bug significantly, in a very low-risk way,
but it doesn't eliminate the bug entirely.

Second, this fix increments the sequence number by 16 on assembly.  This part
of the fix effectively eliminates the possibility that a few accidental
advancements of the sequence number on the failed drive can later put it in a
position to be accepted back into the array.  Like the pre-flight check, this
is not a complete fix, but with both of these in place, it's pretty close.
With this fix, even if the failure originally occurs on cm1 after the
pre-flight check, this advance-by-16 technique ensures that later cm1 is not
able to assemble with its local drive.  It has to wait for cm2 to come back.